Enable EmojiApi test case for latest server

+++ mattermost4j-core/src/test/java/net/bis5/mattermost/client4/MattermostApiTest.java
@@ -1726,7 +1726,6 @@ public class MattermostApiTest {
   // Emoji
 
   @Test
-  @Ignore // emoji upload fail from MM 4.8
   public void testEmoji_CreateCustomEmoji() throws URISyntaxException {
     Path image = Paths.get(getClass().getResource("/noto-emoji_u1f310.png").toURI());
     Emoji emoji = new Emoji();
@@ -1734,21 +1733,29 @@ public class MattermostApiTest {
     emoji.setName(emojiName);
     emoji.setCreatorId(th.basicUser().getId());
 
-    ApiResponse<Emoji> response = assertNoError(client.createEmoji(emoji, image));
-    Emoji createdEmoji = response.readEntity();
+    ApiResponse<Emoji> response = client.createEmoji(emoji, image);
+    if (isNotSupportVersion("5.4.0", response)) {
+      // CreateEmoji call fail between 4.8 and 5.3
+      return;
+    }
+    Emoji createdEmoji = assertNoError(response).readEntity();
 
     assertThat(createdEmoji.getName(), is(emojiName));
     assertThat(createdEmoji.getId(), is(not(nullValue())));
   }
 
   @Test
-  @Ignore // emoji upload fail from MM 4.8
   public void testEmoji_GetCustomEmojiList() throws URISyntaxException {
     Path image = Paths.get(getClass().getResource("/noto-emoji_u1f310.png").toURI());
     Emoji emoji1 = new Emoji();
     emoji1.setName("custom" + th.newId());
     emoji1.setCreatorId(th.basicUser().getId());
-    emoji1 = client.createEmoji(emoji1, image).readEntity();
+    ApiResponse<Emoji> resp1 = client.createEmoji(emoji1, image);
+    if (isNotSupportVersion("5.4.0", resp1)) {
+      // CreateEmoji call fail between 4.8 and 5.3
+      return;
+    }
+    emoji1 = assertNoError(resp1).readEntity();
     Emoji emoji2 = new Emoji();
     emoji2.setName("custom" + th.newId());
     emoji2.setCreatorId(th.basicUser().getId());
@@ -1762,13 +1769,17 @@ public class MattermostApiTest {
   }
 
   @Test
-  @Ignore // emoji upload fail from MM 4.8
   public void testEmoji_GetCustomEmoji() throws URISyntaxException {
     Path image = Paths.get(getClass().getResource("/noto-emoji_u1f310.png").toURI());
     Emoji emoji = new Emoji();
     emoji.setName("custom" + th.newId());
     emoji.setCreatorId(th.basicUser().getId());
-    emoji = client.createEmoji(emoji, image).readEntity();
+    ApiResponse<Emoji> resp1 = client.createEmoji(emoji, image);
+    if (isNotSupportVersion("5.4.0", resp1)) {
+      // CreateEmoji call fail between 4.8 and 5.3
+      return;
+    }
+    emoji = assertNoError(resp1).readEntity();
     String emojiId = emoji.getId();
 
     ApiResponse<Emoji> response = assertNoError(client.getEmoji(emojiId));
@@ -1778,13 +1789,17 @@ public class MattermostApiTest {
   }
 
   @Test
-  @Ignore // emoji upload fail from MM 4.8
   public void testEmoji_DeleteCustomEmoji() throws URISyntaxException {
     Path image = Paths.get(getClass().getResource("/noto-emoji_u1f310.png").toURI());
     Emoji emoji = new Emoji();
     emoji.setName("custom" + th.newId());
     emoji.setCreatorId(th.basicUser().getId());
-    emoji = assertNoError(client.createEmoji(emoji, image)).readEntity();
+    ApiResponse<Emoji> resp1 = client.createEmoji(emoji, image);
+    if (isNotSupportVersion("5.4.0", resp1)) {
+      // CreateEmoji call fail between 4.8 and 5.3
+      return;
+    }
+    emoji = assertNoError(resp1).readEntity();
     String emojiId = emoji.getId();
 
     ApiResponse<Boolean> response = assertNoError(client.deleteEmoji(emojiId));
